Feature 03

See merge request mountblue/29.2-java/bhuvanesh.hj/stackoverflow-clone-grp/stackoverflow-clone!6

+++ src/main/java/com/stackoverflow/dto/TagDTO.java
@@ -9,5 +9,5 @@ import lombok.*;
 @NoArgsConstructor
 public class TagDTO {
     private String name;
-    private Long questionCount;
+    private int questionCount;
 }
